Fix (CICD): building without values

+++ trombone/get_apk.bash
@@ -3,10 +3,14 @@
 set -e
 set -x
 
-cp ./.env.default ./.env
-sed -i "s/=XGROUPAUTHORIZATION/$XGROUPAUTHORIZATION/g" ./.env
-sed -i "s/=API_URL/$API_URL/g" ./.env
+if [ ! -f './.env' ]; then
+    cp ./.env.default ./.env
+    sed -i "s/=XGROUPAUTHORIZATION/=$XGROUPAUTHORIZATION/g" ./.env
+    sed -i "s/=API_URL/=$API_URL/g" ./.env
+fi
 
-cd android
+cd android || exit 1
+
 ./gradlew --no-daemon assembleRelease
+
 cp app/build/outputs/apk/release/app-release.apk ..
